Cleaning up and reusing ShapeRenderer
